tools/model-reader/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "model-reader"
version = "0.1.0"
description = "Read .3dm Rhino files accessibly — explore layers, extract plans, sections, and elevations"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Radical Accessibility Project" },
]
dependencies = [
    "rhino3dm>=8.0.0",
    "click>=8.1.0",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
piaf = ["tactile-core>=1.0.0"]
mcp = ["mcp>=1.0.0"]
dev = ["pytest>=7.0.0", "pytest-cov>=4.0.0"]

[project.scripts]
model-reader = "model_reader.cli:main"
model-reader-mcp = "model_reader.mcp_server.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/model_reader"]

[tool.hatch.build.targets.sdist]
include = ["src/model_reader"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
